ChatBot/requirements.txt
anthropic==1.0.0
python-dotenv==1.0.0

#pip install -r requirements.txt (run in terminal)